refactor: narrow migration counters from u16 to u8
Amp-Thread-ID: https://ampcode.com/threads/T-019cf822-c0f4-70b0-b972-341e4f4fd9ec

+++ crates/precompiles/src/validator_config_v2/mod.rs
@@ -37,8 +37,8 @@ struct Config {
     init_at_height: u64,
     /// Number of V1 validators skipped during migration (bad pubkey, duplicate, etc).
     /// Packed alongside `is_init` and `init_at_height` since all are migration lifecycle state.
-    migration_skipped_count: u16,
-    v1_validator_count: u16,
+    migration_skipped_count: u8,
+    v1_validator_count: u8,
 }
 
 impl Config {
@@ -891,7 +891,7 @@ impl ValidatorConfigV2 {
             if v1_count == 0 {
                 Err(ValidatorConfigV2Error::empty_v1_validator_set())?
             }
-            config.v1_validator_count = v1_count as u16;
+            config.v1_validator_count = v1_count as u8;
             self.config.write(Config {
                 owner: config.owner,
                 is_init: false,
